requirements.txt
-e git+ssh://git@github.com/kaogilvie/datastore-interface.git@63c326419f0ed84df55f7b73935e33dd38748c99#egg=ampersand_datastore
asn1crypto==1.4.0
cachetools==5.3.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-bigquery==3.12.0
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
idna==3.3
oscrypto==1.2.1
packaging==23.2
proto-plus==1.22.3
protobuf==4.24.4
psycopg2-binary==2.9.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.14.1
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
rsa==4.9
six==1.16.0
snowflake-connector-python==2.7.4
urllib3==1.26.8
